Fix the missing account deletion confirmation button
Split trip-index JS file into trip-card and delete-modal.

Move delete modal partial into the layout directory - it is used for
account deletion in addition to trip deletion.
