+++ apps/concierge_site/assets/js/delete-modal.js
@@ -7,6 +7,7 @@ export default $ => {
     const tripId = $(e.relatedTarget).data("trip-id");
     const token = $(e.relatedTarget).data("token");
     if (!tripId) {
+      // The confirmation button has already been set up in the _delete_modal partial
       return;
     }
 
@@ -23,33 +24,14 @@ export default $ => {
       );
   });
 
-  // make entire trip card clickable
-  $("div[data-trip-card='link']").on("click", function(e) {
-    // bail out if the delete modal link is clicked
-    const isModal = $(e.target).data("toggle") == "modal" ? true : false;
-    if (isModal) {
-      return;
-    }
-
-    const $targetEl = $(e.target);
-    const $parentEl = $($targetEl).parents("div[data-trip-card='link']");
-    const linkType = $parentEl.data("link-type");
-    const tripId = $parentEl.data("trip-id");
-    if (!tripId) {
-      return;
-    }
-    const urlBase = (linkType == "accessibility") ? "/accessibility_trips" : "/trips";
-    window.location = `${urlBase}/${tripId}/edit`;
-  });
-
   // make delete modal / links accessible
   $("a[data-toggle='modal']").on("keypress", triggerClick());
-};
 
-function triggerClick() {
-  return function (e) {
-    e.preventDefault();
-    e.stopPropagation();
-    $(e.target).click();
-  };
-}
+  function triggerClick() {
+    return function(e) {
+      e.preventDefault();
+      e.stopPropagation();
+      $(e.target).click();
+    };
+  }
+};
