+++ apps/concierge_site/lib/templates/v2/layout/_delete_modal.html.eex
@@ -3,11 +3,23 @@
     <div class="modal-content">
       <div class="modal-body text-center">
         <div class="my-2">
-          Are you sure you want to delete this subscription?
+          Are you sure you want to delete
+          <%= if assigns[:conn].request_path == "/account/edit" do %>
+            your account?
+          <% else %>
+            this subscription?
+          <% end %>
         </div>
         <div class="my-3" data-modal="action_container">
+          <%# Editing an individual trip %>
           <%= if assigns[:trip] do %>
             <%= link to: v2_trip_path(@conn, :delete, assigns[:trip]), method: :delete, class: "btn btn-primary" do %>
+              Yes, delete
+            <% end %>
+          <% end %>
+          <%# Editing an account %>
+          <%= if assigns[:conn].request_path == "/account/edit" do %>
+            <%= link to: v2_account_path(@conn, :delete), method: :delete, class: "btn btn-primary" do %>
               Yes, delete
             <% end %>
           <% end %>
