+++ apps/concierge_site/lib/templates/v2/trip/index.html.eex
@@ -19,4 +19,4 @@
   </div>
 </div>
 
-<%= render ConciergeSite.V2.TripView, "_delete_modal.html", assigns %>
+<%= render ConciergeSite.V2.LayoutView, "_delete_modal.html", assigns %>
